+++ examples/main/shared/editor/Editor.jsx
@@ -8,7 +8,6 @@ import * as Plugins from './EditorPlugins';
 import ModalsMap from './ModalsMap';
 import theme from '../theme/theme'; // must import after custom styles
 import { GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
-
 const modalStyleDefaults = {
   content: {
     top: '50%',
@@ -51,9 +50,9 @@ export default class Editor extends PureComponent {
           height: testItem.metadata.height,
         };
         setTimeout(() => {
-          updateEntity({ data, files });
+          updateEntity({ data, files /*error: { msg: 'oops :)' }*/ });
           console.log('consumer uploaded', data);
-        }, 500);
+        }, 2000);
       }
     };
     this.helpers = {
@@ -86,6 +85,7 @@ export default class Editor extends PureComponent {
         }, 500);
       },
       onVideoSelected: (url, updateEntity) => {
+        //todo should be moved to videoConfig (breaking change)
         setTimeout(() => {
           const mockVideoIndex =
             this.props.mockImageIndex || Math.floor(Math.random() * testVideos.length);