+++ packages/common/web/src/lib/contentStateServices.js
@@ -1,7 +1,13 @@
-export const truncateContentState = (contentState, index) => {
-  const newContentState = {
-    ...contentState,
-    blocks: [...contentState.blocks.slice(0, index)],
-  };
-  return newContentState;
+import { convertFromRaw, convertToRaw, ContentState } from 'wix-rich-content-editor-common';
+
+export const truncateContentState = (raw, index) => {
+  const contentState = convertFromRaw(raw);
+  const blocks = contentState.getBlocksAsArray();
+  if (index < 0 || index > blocks.length) {
+    return raw;
+  }
+  const truncateBlocks = blocks.slice(0, index);
+  const newContentState = ContentState.createFromBlockArray(truncateBlocks);
+  const newRaw = convertToRaw(newContentState);
+  return newRaw;
 };