+++ packages/common/web/src/test/index.js
@@ -1,0 +1,6 @@
+import raw from './fixtures/raw.json';
+import emptyRaw from './fixtures/emptyRaw.json';
+import expectedRaw1 from './fixtures/expectedRaw1.json';
+import expectedRaw2 from './fixtures/expectedRaw2.json';
+
+export { raw, emptyRaw, expectedRaw1, expectedRaw2 };