+++ CHANGELOG.md
@@ -15,26 +15,47 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `code-block`
+  - [#569](https://github.com/wix-incubator/rich-content/pull/569) add code block style, more information in: https://github.com/wix-incubator/rich-content/blob/master/docs/PluginCustomization.md
+
+### :house: Internal
+- `wrapper`
+  - [#980](https://github.com/wix-incubator/rich-content/pull/980) hotfix: createEmpty import
+- `general`
+  - [#985](https://github.com/wix-incubator/rich-content/pull/985) enable publishing with custom npm tag
+
+</details>
+<hr/>
+
+## 7.4.5 (Apr 28, 2020)
 ### :rocket: New Feature
 - `button`
-  - [#953](https://github.com/wix-incubator/rich-content/pull/953) action button: by passing onClick callback, on click triggers it
+  - [#953](https://github.com/wix-incubator/rich-content/pull/953) action button - button with onClick callback
+- `image` `video`
+  - [#972](https://github.com/wix-incubator/rich-content/pull/972) handle upload error - show message on block
 ### :bug: Bug Fix
 - `editor`
   - [#951](https://github.com/wix-incubator/rich-content/pull/951) fix: tooltip opacity
   - [#957](https://github.com/wix-incubator/rich-content/pull/957) fix: command+ctrl+j creates code block on mac
 - `map`
   - [#959](https://github.com/wix-incubator/rich-content/pull/959) modal settings fixed (convention) & made compatible with wrapper palette colors
-- `code-block`
-  - [#569](https://github.com/wix-incubator/rich-content/pull/569) add code block style, more information in: https://github.com/wix-incubator/rich-content/blob/master/docs/PluginCustomization.md
+- `emoji`
+  - [#973](https://github.com/wix-incubator/rich-content/pull/973) was using old editor and editor-common dependencies
+- `video`
+  - [#974](https://github.com/wix-incubator/rich-content/pull/960) video file upload not working when block isn't selected
+- `gallery`
+  - [#963](https://github.com/wix-incubator/rich-content/pull/963) fix:Gallery doesn't show more then 3 images on load 
 ### :house: Internal
 - `editor`
   - [#936](https://github.com/wix-incubator/rich-content/pull/936) arrangement of inline toolbar buttons
 - `wrapper`
   - [#956](https://github.com/wix-incubator/rich-content/pull/956) bi getData will now contain `forPublish` argument
+  - [#966](https://github.com/wix-incubator/rich-content/pull/966) passing RCE/RCV to wrapper is no longer required
+  - [#975](https://github.com/wix-incubator/rich-content/pull/975) ssr fix - render suspense after mount
+  - [#977](https://github.com/wix-incubator/rich-content/pull/977) wrapper exports its type declarations
 - `preview`
   - [#962](https://github.com/wix-incubator/rich-content/pull/962) interactions improved; read-more displays html
-</details>
-<hr/>
 
 ## 7.3.5 (Apr 21, 2020)
 ### :bug: Bug Fix